+++ pytest_common_subject/__init__.py
@@ -2,7 +2,11 @@ import pkg_resources
 
 from .exceptions import DeferredCommonSubjectRvalUsage
 from .fixtures import precondition_fixture
-from .mixins import CommonSubjectTestMixin, WithCommonSubjectDeferred
+from .mixins import (
+    CommonSubjectTestMixin,
+    AsyncCommonSubjectTestMixin,
+    WithCommonSubjectDeferred,
+)
 
 __version__ = pkg_resources.get_distribution('pytest-common-subject').version
 
@@ -10,5 +14,6 @@ __all__ = [
     'DeferredCommonSubjectRvalUsage',
     'precondition_fixture',
     'CommonSubjectTestMixin',
+    'AsyncCommonSubjectTestMixin',
     'WithCommonSubjectDeferred',
 ]